fix: 修复导入其他实例配置无法全部导入的 bug

+++ src/BedrockBoot/Views/DialogContent/DialogMigrationGameConfigContent.axaml.cs
@@ -28,7 +28,7 @@ public partial class DialogMigrationGameConfigContent : UserControl
                     Dispatcher.UIThread.Invoke(() =>
                     {
                         if (MigrationProgressBar.IsIndeterminate) MigrationProgressBar.IsIndeterminate = false;
-                        MigrationProgressBar.Value = progress.Percentage;
+                        MigrationProgressBar.Value = (int)progress.Percentage;
                         MigrationProgressText.Text = $"迁移中... ({progress.Percentage:F2} %)";
                     });
                 })
